automation-service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
py-eureka-client